Define NTSTATUS for hidpi.h, which no user-mode header does
hidpi.h declares every HidP_ function as returning NTSTATUS and leaves the
type to somebody else. windows.h does not provide it and the header that
would is a driver header, so MSVC read a hundred prototypes as implicit-int
variables and reported the last as a redefinition of the first.

One identical redeclaration, which is legal in C and C++ alike — so it stays
correct on a toolchain whose headers do define it, and the cross-check that
found the rest of this now runs with no compatibility flags of its own.

With it, the conversions /WX would have caught next: the report buffers fill
from an unsigned int and a plain 0, and two file sizes compare a DWORD
against a size_t.

+++ helpers/windows/src/secret_store.cpp
@@ -40,7 +40,7 @@ std::vector<uint8_t> read_file(const std::wstring &path) {
         bytes.resize(static_cast<size_t>(size.QuadPart));
         DWORD read = 0;
         if (!ReadFile(file, bytes.data(), static_cast<DWORD>(bytes.size()), &read, nullptr) ||
-            read != bytes.size())
+            read != static_cast<DWORD>(bytes.size()))
             bytes.clear();
     }
     CloseHandle(file);
@@ -59,7 +59,7 @@ bool write_file(const std::wstring &path, const std::vector<uint8_t> &bytes) {
     DWORD written = 0;
     const bool ok = WriteFile(file, bytes.data(), static_cast<DWORD>(bytes.size()), &written,
                               nullptr) &&
-                    written == bytes.size();
+                    written == static_cast<DWORD>(bytes.size());
     CloseHandle(file);
     if (!ok) {
         DeleteFileW(temporary.c_str());
